Fix #57 :: minor tweaks to URL regex
The URL regex used to modify `location.href` to generate the
baseUrl for JS and CSS assets was stripping off the project
name if it contained 'index'.

Modified `index[^\.]*` to `index[^\./]*` so it doesn't match
in the middle of the path.

See [this regexr to view the problem](http://regexr.com/3bjqc)